start benchmark (does NOT work!!!)
